third_party/blink/web_tests/platform/generic/external/wpt/css/css-tables/html-to-css-mapping-2-expected.txt

This is a testharness.js-based test.
PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
PASS 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
FAIL 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style) assert_equals: expected "outset" but got "solid"
FAIL 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color) assert_equals: expected "rgb(0, 0, 0)" but got "rgb(128, 128, 128)"
FAIL 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width) assert_equals: expected "10px" but got "1px"
FAIL 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style) assert_equals: expected "outset" but got "solid"
PASS 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top width) assert_equals: expected "1px" but got "0px"
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top style) assert_equals: expected "inset" but got "none"
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top color) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 0)"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left width) assert_equals: expected "1px" but got "0px"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left style) assert_equals: expected "inset" but got "none"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left color) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 0)"
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left width)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left style)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left color)
PASS 5a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
FAIL 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right width) assert_equals: expected "10px" but got "1px"
PASS 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right style)
PASS 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right color)
Harness: the test ran to completion.
